Allow double-precision sized constant float expressions.  Resolves #1818
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@23257 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/avt/Expressions/General/avtConstantFunctionExpression.h
@@ -67,6 +67,9 @@ class     vtkDataArray;
 //    Jeremy Meredith, Wed Feb 20 10:00:31 EST 2008
 //    Support either nodal or zonal values.
 //
+//    Kathleen Biagas, Thu May  1 17:18:30 PDT 2014
+//    Changed value to double.
+//
 // ****************************************************************************
 
 class EXPRESSION_API avtConstantFunctionExpression
@@ -86,7 +89,7 @@ class EXPRESSION_API avtConstantFunctionExpression
     virtual int           NumVariableArguments(void) { return 2; }
   private:
     bool    nodal;
-    float   value;
+    double  value;
 };
